Files added
lights_mqtt.py - Main script that listens to an mqtt topic and executes functions when key words are seen, like "green" will send a serial command to the light tower turn on the green light. Or, 'flash' to execute a series of serial commands to make various lights on the tower flash. There are other functions, like locking a windows desktop and googles text to speech that can be removed or expanded.
randomness.py - a sub script that executes when the key word 'random' is seen. It scrapes reddit's 'shower thoughts' and sends back into the topic a random 'shower thought' (NSFW)
new_lights.html and new_server.js - these allow you to send text/keywords to the mqtt topic over websockets. Also included is a javascript visualization of which color the light tower might be showing based on the last key word sent to the mqtt topic. You can see the text in a text box at the bottom of the page too. Everyone with the html file open will see the same text and the color visualization will be synchronized across all computers/browser tabs. 
mosquitto.org.crt - this is needed to connect to the test MQTT mosquitto server, not needed if you're going to host your own mosquitto server with it's own SSL cert.